+++ src/array.rs
@@ -76,6 +76,60 @@ where
 }
 
 // ---------------------------------------------------------------------------------------
+// FoldAdd
+
+/// Hide our `Null` type
+const _: () = {
+    /// A type which contributes nothing when adding (i.e. a zero)
+    pub struct Null;
+    impl<T> Add<T> for Null {
+        type Output = T;
+        fn add(self, rhs: T) -> Self::Output {
+            rhs
+        }
+    }
+
+    impl FoldAdd for ATerm {
+        type Output = Null;
+    }
+};
+
+impl<V, A> FoldAdd for TArr<V, A>
+where
+    A: FoldAdd,
+    FoldSum<A>: Add<V>,
+{
+    type Output = Sum<FoldSum<A>, V>;
+}
+
+// ---------------------------------------------------------------------------------------
+// FoldMul
+
+/// Hide our `Null` type
+const _: () = {
+    /// A type which contributes nothing when multiplying (i.e. a one)
+    pub struct Null;
+    impl<T> Mul<T> for Null {
+        type Output = T;
+        fn mul(self, rhs: T) -> Self::Output {
+            rhs
+        }
+    }
+
+    impl FoldMul for ATerm {
+        type Output = Null;
+    }
+};
+
+impl<V, A> FoldMul for TArr<V, A>
+where
+    A: FoldMul,
+    FoldProd<A>: Mul<V>,
+{
+    type Output = Prod<FoldProd<A>, V>;
+}
+
+// ---------------------------------------------------------------------------------------
 // Add arrays
 // Note that two arrays are only addable if they are the same length.